requirements.txt
python-dotenv
tweepy
transformers
torch
jax
tensorflow
fugashi
ipadic